requirements.txt
pyrogram==2.0.106
TgCrypto==1.2.5
motor==3.3.2
pymongo==4.3.3
pytz==2023.3
requests==2.31.0
aiohttp==3.9.1
python-dotenv==1.0.0
